moved config files to config folder, added config folder to .gitignore and update imports
